Now I can show the wishlist products

+++ resources/views/products/wishlist.blade.php
@@ -7,22 +7,24 @@
 
           @foreach ($products as $product)
           <div class="card">
-              <p class="card-text">{{$product->name}}</p>
+            @if(!is_null($product['name']))
+              <p class="card-text">{{$product['name']}}</p>
               <ul>
-                <li>{{$product->colour}}</li>
-                <li>{{$product->type}}</li>
-                <li>{{$product->brand}}</li>
+                <li>{{$product['colour']}}</li>
+                <li>{{$product['type']}}</li>
+                <li>{{$product['brand']}}</li>
               </ul>
               <p style="text-align: right; color: red; font-size: 3rem; ">
-                €{{$product->price}}
+                €{{$product['price']}}
               </p>
               @if(auth()->check())
                 <form name="delete_from_wishlist" method="DELETE" action="/wishlist/delete">
                   {{ csrf_field() }}
-                  <input type="hidden" name="product_id" value="{{$product->id}}" />
+                  <input type="hidden" name="product_id" value="{{$product['id']}}" />
                   <button type="submit" class="btn btn-danger" style="width: 100%;">Delete</button>
                 </form>
               @endif
+            @endif
           </div>
           @endforeach
         </div>
